updates unittests for FileInput and FileInputRenderer
