+++ src/signals/incident/components/form/FileInput/style.scss
@@ -7,12 +7,8 @@ $iconLoadingSize: 40px;
 
 
 .file-input {
-  display: flex;
-  height: calc(#{$boxSize} + (2 * #{$borderWidth}));
 
   &__preview {
-    width: calc(#{$boxSize} + (2 * #{$borderWidth}));
-    margin-right: 10px;
 
     &--loading {
       border: 1px solid #bebebe;
@@ -52,15 +48,6 @@ $iconLoadingSize: 40px;
   }
 
   &__button {
-    border: 1px dashed #bebebe;
-    width: calc(#{$boxSize} + (2 * #{$borderWidth}));
-    margin-right: 10px;
-
-    input[type=file] {
-      opacity: 0;
-      width: 0;
-      height: 0;
-    }
 
     &-label {
       position: relative;
@@ -90,9 +77,4 @@ $iconLoadingSize: 40px;
     }
   }
 
-  &__empty {
-    width: calc(#{$boxSize} + (2 * #{$borderWidth}));
-    border: 1px dashed #bebebe;
-    margin-right: 10px;
-  }
 }
